Move also TypedDict
