+++ typing_extensions/src_py2/test_typing_extensions.py
@@ -749,85 +749,6 @@ class ProtocolTests(BaseTestCase):
             self.assertIsInstance(E(), D)
 
 
-class TypedDictTests(BaseTestCase):
-
-    def test_basics_iterable_syntax(self):
-        Emp = TypedDict('Emp', {'name': str, 'id': int})
-        self.assertIsSubclass(Emp, dict)
-        self.assertIsSubclass(Emp, typing.MutableMapping)
-        if sys.version_info[0] >= 3:
-            import collections.abc
-            self.assertNotIsSubclass(Emp, collections.abc.Sequence)
-        jim = Emp(name='Jim', id=1)
-        self.assertIs(type(jim), dict)
-        self.assertEqual(jim['name'], 'Jim')
-        self.assertEqual(jim['id'], 1)
-        self.assertEqual(Emp.__name__, 'Emp')
-        self.assertEqual(Emp.__module__, __name__)
-        self.assertEqual(Emp.__bases__, (dict,))
-        self.assertEqual(Emp.__annotations__, {'name': str, 'id': int})
-        self.assertEqual(Emp.__total__, True)
-
-    def test_basics_keywords_syntax(self):
-        Emp = TypedDict('Emp', name=str, id=int)
-        self.assertIsSubclass(Emp, dict)
-        self.assertIsSubclass(Emp, typing.MutableMapping)
-        if sys.version_info[0] >= 3:
-            import collections.abc
-            self.assertNotIsSubclass(Emp, collections.abc.Sequence)
-        jim = Emp(name='Jim', id=1)
-        self.assertIs(type(jim), dict)
-        self.assertEqual(jim['name'], 'Jim')
-        self.assertEqual(jim['id'], 1)
-        self.assertEqual(Emp.__name__, 'Emp')
-        self.assertEqual(Emp.__module__, __name__)
-        self.assertEqual(Emp.__bases__, (dict,))
-        self.assertEqual(Emp.__annotations__, {'name': str, 'id': int})
-        self.assertEqual(Emp.__total__, True)
-
-    def test_typeddict_errors(self):
-        Emp = TypedDict('Emp', {'name': str, 'id': int})
-        self.assertEqual(TypedDict.__module__, 'typing_extensions')
-        jim = Emp(name='Jim', id=1)
-        with self.assertRaises(TypeError):
-            isinstance({}, Emp)
-        with self.assertRaises(TypeError):
-            isinstance(jim, Emp)
-        with self.assertRaises(TypeError):
-            issubclass(dict, Emp)
-        with self.assertRaises(TypeError):
-            TypedDict('Hi', x=1)
-        with self.assertRaises(TypeError):
-            TypedDict('Hi', [('x', int), ('y', 1)])
-        with self.assertRaises(TypeError):
-            TypedDict('Hi', [('x', int)], y=int)
-
-    def test_pickle(self):
-        global EmpD  # pickle wants to reference the class by name
-        EmpD = TypedDict('EmpD', name=str, id=int)
-        jane = EmpD({'name': 'jane', 'id': 37})
-        for proto in range(pickle.HIGHEST_PROTOCOL + 1):
-            z = pickle.dumps(jane, proto)
-            jane2 = pickle.loads(z)
-            self.assertEqual(jane2, jane)
-            self.assertEqual(jane2, {'name': 'jane', 'id': 37})
-            ZZ = pickle.dumps(EmpD, proto)
-            EmpDnew = pickle.loads(ZZ)
-            self.assertEqual(EmpDnew({'name': 'jane', 'id': 37}), jane)
-
-    def test_optional(self):
-        EmpD = TypedDict('EmpD', name=str, id=int)
-
-        self.assertEqual(typing.Optional[EmpD], typing.Union[None, EmpD])
-        self.assertNotEqual(typing.List[EmpD], typing.Tuple[EmpD])
-
-    def test_total(self):
-        D = TypedDict('D', {'x': int}, total=False)
-        self.assertEqual(D(), {})
-        self.assertEqual(D(x=1), {'x': 1})
-        self.assertEqual(D.__total__, False)
-
-
 class AnnotatedTests(BaseTestCase):
 
     def test_repr(self):
